improve tracking semantics
- om.runtime.model('foo').experiment() returns a tracker
- om.runtime.model('foo').experiments() returns all trackers
- enable backend.predict calls to record actual X, Y values
- refactor tracking backend to easy extensions
- update correctness summary() for multiple metrics
