+++ omegaml/backends/tensorflow/tfsavedmodel.py
@@ -144,7 +144,7 @@ class TensorflowSavedModelBackend(BaseModelBackend):
 
         """
         model = self.get_model(modelname)
-        X = self.data_store.get(Xname)
+        X = self._resolve_input_data('predict', Xname, **kwargs)
         result = model.predict(X)
 
         def ensure_serializable(data):
@@ -159,10 +159,7 @@ class TensorflowSavedModelBackend(BaseModelBackend):
             return data
 
         result = ensure_serializable(result)
-
-        if rName:
-            result = self.data_store.put(result, rName)
-        return result
+        return self._prepare_result('predict', result, rName=rName, pure_python=pure_python, **kwargs)
 
     def fit(self, modelname, Xname, Yname=None, pure_python=True, tpu_specs=None, **kwargs):
         raise ValueError('cannot fit a saved model')
