+++ omegaml/runtimes/modelproxy.py
@@ -52,6 +52,9 @@ class OmegaModelProxy(object):
         self.apply_mixins()
         self.apply_require()
 
+    def __repr__(self):
+        return f'OmegaModelProxy({self.modelname})'
+
     def apply_require(self):
         self.meta = meta = self.runtime.omega.models.metadata(self.modelname)
         assert meta is not None, "model {self.modelname} does not exist".format(**locals())
@@ -80,3 +83,47 @@ class OmegaModelProxy(object):
         return the task from the runtime with requirements applied
         """
         return self.runtime.task(name)
+
+    def experiment(self, experiment=None, label=None, provider=None):
+        """ return the experiment for this model
+
+        If an experiment does not exist yet, it will be created. The
+        experiment is automatically set to track this model, unless another
+        model has already been set to track this model for the same label.
+        If a previous model has been set to track this model it will be
+        returned. If an experiment name is passed it will be used.
+
+        Args:
+            experiment (str): the experiment name, defaults to the modelname
+            label (str): the runtime label, defaults to 'default'
+            provider (str): the provider to use, defaults to 'default'
+
+        Returns:
+            OmegaTrackingProxy() instance
+        """
+        label = label or self.runtime._default_label
+        exps = self.experiments(label=label) if experiment is None else None
+        exp = exps[0] if exps else None
+        experiment = experiment or self.modelname
+        if exp is None:
+            exp = self.runtime.experiment(experiment, provider=provider)
+            if not label in self.experiments():
+                exp.track(self.modelname, label=label)
+        return exp
+
+    def experiments(self, label=None, raw=False):
+        """ return list of experiments tracking this model
+
+        Args:
+            label (None|str): if set, return only the experiment for this label
+            raw (bool): if True return the metadata for the experiment, else return the OmegaTrackingProxy
+
+        Returns:
+            list of OmegaTrackingProxy instances or Metadata objects
+        """
+        store = self.runtime.omega.models
+        tracking = (store.metadata(self.modelname).attributes.get('tracking', {}))
+        names = [tracking.get(label)] if label in tracking else None
+        names = names or (tracking.get('experiments', []) if not (label or names) else [])
+        return [self.runtime.experiment(name) if not raw else store.metadata(f'experiments/{name}')
+                for name in names]
